+++ modules/core.lua
@@ -1,7 +1,7 @@
 local mod = ...
 
 local M = {
-  version = "2.1.0",
+  version = "2.1.1",
   game = nil,
   linkBattleActive = false,
   startItems = {},